Just post creator can edit and delete its posts

+++ app/Http/Controllers/PostsController.php
@@ -86,7 +86,14 @@ class PostsController extends Controller
      */
     public function edit($id)
     {
+
         $post  =Post::find($id);
+
+        //Check for correct user
+        if(auth()->user()->id != $post->user_id)
+        {
+          return redirect('/posts')->with('errors','Unauthoerisxe user');
+        }
         return view('posts.edit')->with('post', $post);
     }
 
@@ -122,6 +129,11 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+        //Check for correct user
+        if(auth()->user()->id != $post->user_id)
+        {
+          return redirect('/posts')->with('errors','Unauthoerisxe user');
+        }
         $post->delete();
         return redirect('/posts')->with('success','Posts deleted!');
     }
